Added Art Folders and Singleton
Added art folders for character sprites for the artists to populate. Will also import some of the old art to there soon. Also updated GameManager and RelationshipFramework to be singleton. Will need to do more work on it though.

+++ Assets/scripts/GameManagers/DataPersistence/GameData.cs
@@ -24,13 +24,13 @@ public class GameData
     public List<Buff> activeBuffs;
     public List<Debuff> activeDebuffs;
 
-    public float copRelations;
-    public float russianMobRelations;
-    public float norwegianMobRelations;
-    public float sicilianMobRelations;
-    public float prohibitionistsRelations;
-    public float drunkardRelations;
-    public float civilianRelations;
+    public int copRelations;
+    public int russianMobRelations;
+    public int norwegianMobRelations;
+    public int sicilianMobRelations;
+    public int prohibitionistsRelations;
+    public int drunkardRelations;
+    public int civilianRelations;
 
     public int cash;
     public int paneling;
